feat: modify domain features

+++ backend/src/main/java/com/preproject/backend/domain/question/dto/QuestionDto.java
@@ -39,7 +39,7 @@ public class QuestionDto {
     @Setter
     @AllArgsConstructor
     public static class Response {
-        private int questionId;
+        private long questionId;
         private String title;
         private String content;
         private LocalDateTime createdAt;
